Add platform icons to GameCard
A new component, PlatformIconList, has been created to display platform-specific icons on the GameCard. The "react-icons" package was installed to facilitate this, and a "Platform" interface has been added to "useGames.tsx" to handle platform data. This will improve the visual representation of each game's supported platforms, contributing to better user experience.