Replace GameModal with ModalGame component
The GameModal component has been replaced with a new component called ModalGame in the GameCard component. This new ModalGame component accepts a game id as a prop and handles the functionality previously found in GameModal. The ModalGame component display detailed information about the game when the "Show More" button is clicked.